requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
pytest
httpx
pytest-playwright
jinja2
email-validator
python-jose[cryptography]
python-multipart
bcrypt